fix: rename indeminite->indemnite
