+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/ResultDetail.js
