+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/Salaire.js
