+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/SimpleSwitch.js
