+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/indemnite.test.js
